+++ examples/scrape-example.yaml
@@ -1,0 +1,15 @@
+scrape_webpage:
+    input: https://example.com
+    model: gpt-4o-mini
+    action: Analyze the scraped content and provide insights
+    output: STDOUT
+    scrape_config:
+        allowed_domains:
+            - example.com
+        headers:
+            Accept: text/html,application/xhtml+xml
+            Accept-Language: en-US,en;q=0.9
+        extract:
+            - title
+            - text
+            - links
